pyproject.toml
[project]
name = "kdocs-synth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.2",
    "opencv-contrib-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pyyaml>=6.0.2",
]
